get back to float

+++ Classes/TimeFeatures.cpp
@@ -16,7 +16,7 @@ TimeFeatures::~TimeFeatures(){
   
 }
 
-vector<double> TimeFeatures::getFeatures(vector<vector<double> > data)
+vector<float> TimeFeatures::getFeatures(vector<vector<float> > data)
 {
   featList.clear();
   for (int i = 0; i<data.size(); i++)   // calculate the features per dimension
@@ -32,49 +32,49 @@ vector<double> TimeFeatures::getFeatures(vector<vector<double> > data)
   return featList;
 }
 
-void TimeFeatures::normalization(vector<double>& data)
+void TimeFeatures::normalization(vector<float>& data)
 {
-  double ref=getMax(data);
+  float ref=getMax(data);
   for (int i = 0; i<data.size(); i++)
     data[i] = data[i]/ref;
 }
 
-void TimeFeatures::featZcr(vector<double>& data)
+void TimeFeatures::featZcr(vector<float>& data)
 {
   // since this is the first feature we calculate, we clear the featlist
   size_t numSamples = data.size();
-  vector<double> ZCRList;
+  vector<float> ZCRList;
   for (size_t i = 0; i<numSamples; i=i+hopSize)
   {
     if (i+blockSize-1>=numSamples)
       break;
     else
     {
-      vector<double>::const_iterator first = data.begin()+i;
-      vector<double>::const_iterator last = data.begin()+i+blockSize;
-      vector<double> temp(first,last);
+      vector<float>::const_iterator first = data.begin()+i;
+      vector<float>::const_iterator last = data.begin()+i+blockSize;
+      vector<float> temp(first,last);
       int zcrCount = 0;
       for (int j =1; j<blockSize; j++)
       {
         if (temp[j]*temp[j-1]<0)
           zcrCount+=1;
       }
-      ZCRList.push_back((double)zcrCount/blockSize);
+      ZCRList.push_back((float)zcrCount/blockSize);
     }
   }
-  double meanZCR = getSum(ZCRList)/ZCRList.size();
-  double sq_sum = 0.0;
+  float meanZCR = getSum(ZCRList)/ZCRList.size();
+  float sq_sum = 0.0;
   for (int j = 0; j<ZCRList.size(); j++)
     sq_sum+=(ZCRList[j]-meanZCR)*(ZCRList[j]-meanZCR);
-  double stdevZCR = sqrt(sq_sum/(float)(ZCRList.size()-1));
+  float stdevZCR = sqrt(sq_sum/(float)(ZCRList.size()-1));
   featList.push_back(meanZCR);
   featList.push_back(stdevZCR);
 }
 
-void TimeFeatures::featRms(vector<double>& data)
+void TimeFeatures::featRms(vector<float>& data)
 {
   size_t numSamples = data.size();
-  vector <double> rmsList;
+  vector <float> rmsList;
   for (size_t i = 0; i<numSamples; i = i+hopSize)
   {
     if (i+blockSize-1>=numSamples)
@@ -82,9 +82,9 @@ void TimeFeatures::featRms(vector<double>& data)
     else
     {
       float sum = 0;
-      vector<double>::const_iterator first = data.begin()+i;
-      vector<double>::const_iterator last = data.begin()+i+blockSize;
-      vector<double> temp(first,last);
+      vector<float>::const_iterator first = data.begin()+i;
+      vector<float>::const_iterator last = data.begin()+i+blockSize;
+      vector<float> temp(first,last);
       for (int j = 0; j<blockSize; j++)
       {
         sum = sum + temp[j]*temp[j];
@@ -92,34 +92,34 @@ void TimeFeatures::featRms(vector<double>& data)
       rmsList.push_back(sqrt(sum/blockSize));
     }
   }
-  double meanRms = getSum(rmsList)/rmsList.size();
-  double sq_sum = 0.0;
+  float meanRms = getSum(rmsList)/rmsList.size();
+  float sq_sum = 0.0;
   for (int j = 0; j<rmsList.size(); j++)
     sq_sum+=(rmsList[j]-meanRms)*(rmsList[j]-meanRms);
-  double stdevRms = sqrt(sq_sum/(float)(rmsList.size()-1));
+  float stdevRms = sqrt(sq_sum/(float)(rmsList.size()-1));
   featList.push_back(meanRms);
   featList.push_back(stdevRms);
 }
 
-void TimeFeatures::featEnv(vector<double>& data)
+void TimeFeatures::featEnv(vector<float>& data)
 {
   size_t numSamples = data.size();
-  vector<double> envList;
+  vector<float> envList;
   for (size_t i = 0 ; i<numSamples; i=i+hopSize)
   {
     if (i+blockSize-1>=numSamples)
       break;
     else
     {
-      vector<double>::const_iterator first = data.begin()+i;
-      vector<double>::const_iterator last = data.begin()+i+blockSize;
-      vector<double> temp(first,last);
+      vector<float>::const_iterator first = data.begin()+i;
+      vector<float>::const_iterator last = data.begin()+i+blockSize;
+      vector<float> temp(first,last);
       float tempMaxEnv = getMax(temp);
       envList.push_back(tempMaxEnv);
     }
   }
-  double meanEnv = getSum(envList)/envList.size();
-  double sq_sum = 0.0;
+  float meanEnv = getSum(envList)/envList.size();
+  float sq_sum = 0.0;
   for (int j = 0; j<envList.size(); j++)
     sq_sum+=(envList[j]-meanEnv)*(envList[j]-meanEnv);
   float stdevEnv = sqrt(sq_sum/(float)(envList.size()-1));
@@ -127,23 +127,23 @@ void TimeFeatures::featEnv(vector<double>& data)
   featList.push_back(stdevEnv);
 }
 
-void TimeFeatures::featJerk(vector<double>& data)
+void TimeFeatures::featJerk(vector<float>& data)
 {
   
 }
 
-void TimeFeatures :: featSkewAndKurt(vector<double>& data)
+void TimeFeatures :: featSkewAndKurt(vector<float>& data)
 {
   size_t numSamples = data.size();
-  double sum = getSum(data);
-  double mean = sum / numSamples;
+  float sum = getSum(data);
+  float mean = sum / numSamples;
   
   float sq_sum = 0;
   for (size_t i=0; i<numSamples; i++)
     sq_sum += (data[i] - mean)*(data[i] - mean);
-  double stdev = sqrt(sq_sum / (data.size()-1));
+  float stdev = sqrt(sq_sum / (data.size()-1));
   
-  double skewness = 0;
+  float skewness = 0;
   for (size_t i = 0; i < numSamples; i++)
     skewness += (data[i] - mean)*(data[i] - mean)*(data[i] - mean);
   skewness = skewness/(numSamples * stdev * stdev * stdev);
@@ -156,7 +156,7 @@ void TimeFeatures :: featSkewAndKurt(vector<double>& data)
   featList.push_back(kurtosis);
 }
 
-double TimeFeatures::getSum(vector<double>& tempData)
+float TimeFeatures::getSum(vector<float>& tempData)
 {
   float sum = 0;
   for (int i=0; i<tempData.size(); i++)
@@ -164,9 +164,9 @@ double TimeFeatures::getSum(vector<double>& tempData)
   return sum;
 }
 
-double TimeFeatures::getMax(vector<double>& tempData, int start, int end)
+float TimeFeatures::getMax(vector<float>& tempData, int start, int end)
 {
-  double max = 0;
+  float max = 0;
   if (start >= end)
   {
     for (int i=0; i<tempData.size(); i++)
@@ -186,9 +186,9 @@ double TimeFeatures::getMax(vector<double>& tempData, int start, int end)
   return max;
 }
 
-double TimeFeatures::getMin(vector<double>& tempData, int start, int end)
+float TimeFeatures::getMin(vector<float>& tempData, int start, int end)
 {
-  double min = 10000;
+  float min = 10000;
   if (start>=end)
   {
     for (int i=0; i<tempData.size(); i++)
